requirements.txt
ftfy
regex
tqdm
torch
torchvision
neurosity
python-dotenv
matplotlib
numpy
pandas
torch
pygame
requests
Pillow
